// lib/sdk/src/services/middleware-api/types/openapi-types.ts
export interface components {
  schemas: {
    AgentAvailability: {
      available: boolean;
      /** Estimated wait time in seconds */
      estimatedWaitTime?: number;
      /** Represents when a program is open or closed */
      hoursOfOperation?: boolean;
      /** Number of chats waiting in the queue. */
      queueDepth?: number;
      /** Indicates whether there is an active agent with a chat slot available (online), there are agents logged in, but all their chat slots are full (busy), or all agents on a busy status or are not logged in (offline). */
      status?: string;
    };
    BadRequest: {
      errors: components["schemas"]["Error"][];
      message: string;
    };
    End: {
      senderId: string;
    };
    Error: {
      message: string;
      source: string;
      type: string;
    };
    Escalate: {
      skill: string;
      userId: string;
    };
    EscalateResponse: {
      agentId?: string;
      escalationId: string;
      /** Estimated wait time in seconds */
      estimatedWaitTime?: number;
      /** The user position in the chat queue. */
      queuePosition?: number;
      /** (accepted, queued) */
      status?: string;
    };
    History: {
      messages: components["schemas"]["MessageSent"][];
      pagination: {
        page: number;
        totalPages: number;
      };
    };
    InternalError: {
      message: string;
    };
    Message: {
      content: string;
      /** The user or agent ID */
      senderId: string;
      /** Who is sending the message (agent or user) */
      side: string;
    };
    MessageSent: {
      content: string;
      /** The user or agent ID */
      senderId: string;
      sentDate: string;
      /** Who is sending the message (agent or user) */
      side: string;
    };
    Setting: {
      callbackToken: string;
      callbackURL: string;
      integrationFields: { [key: string]: unknown };
      integrationName: string;
    };
    Typing: {
      typing: boolean;
    };
    WaitTime: {
      /** Estimated wait time in seconds */
      estimatedWaitTime: number;
    };
  };
  responses: {};
}